SMAGENT-1746: Add REST endpoint for latest metrics (#1470)
* SMAGENT-1746: Add REST endpoint for latest metrics

This change adds a new REST endpoint `/api{,/v0}/protobuf/metrics` that
returns the latest complete protobuf in JSON format.

I change the metric_serlializer to store the metrics with a
`std::shared_ptr`, so that I can save a copy of that pointer for the
endpoint to read later.  That gets stored within the `metric_store`
namespace.  The `metrics_rest_request_handler` handles the REST request.