SMAGENT-1259: Add Fault Injection Framework (#1475)
* SMAGENT-1259: Add Fault Injection Framework

This change introduces a fault injection framework into the agent.  The
goal of this framework is to enable engineers to programmatically inject
faults into the running system.  This will enable use to test error
paths in the code that are difficult (or near impossible) to trigger.

Note that this change includes only the framework -- there is not way to
actually enable any faults at the point.  That will come in a future
change.