SMAGENT-1637: Avoid "incomplete" containers (#1459)
* SMAGENT-1637: Avoid "incomplete" containers

This change updates the `sinsp_analyzer` to not emit any container whose
metadata is incomplete.  The expectation is that a container's metadata
can be incomplete only between the time we realize that there might be a
new container and the time the async lookup completes.  We should, then,
expect to skip the emission for a flush cycle or maybe two.

* Revert "SMAGENT-1637: Avoid "incomplete" containers"

This reverts commit d82987c88c8c99c67d87c99d424146e39d908599.

* Rework approach

This change reworks the original approach to solving this problem.
Here, if the metadata is incomplete, we still emit the container, we
just don't fill in the "incomplete" fields.

I tested this by:
* Tweaking the docker metadata fetcher to drop the first N requests
* Run some containers, each of which show up as "n/a"
* Once N requests pass and the docker metadata fetcher starts actually
  satisfying requests, exec a shell into the container
* I observe the container originally marked as "n/a" get their metadata
  fill in correct; I no longer see them in the "n/a" grouping.