Remove fault (#1484)
I was supposed to remove this before I submitted it.  I think I made the
change and forgot to push it to github.

The nice thing is that it does reveal a potential issue with fault
injection points in release builds.  The fault injection points, in
release builds, evaulate to simple things; this one was:

```c++
false;
```

The compiler didn't like that (because of `-Werror`).