Fix typo in infra state. SMAGENT-1776
Discussed this with Mattia, ended up with the conclusion that
this `++j` (or `++i`) is dead code anyway (the node objects
have exactly one parent and it's of the `host` type, so we hit
the `break` every time we enter this loop).

If this code _wasn't_ dead, we'd hang as we never increase `j`
in the inner loop so we would be stuck there forever.