SMAGENT-1777: Fix code coverage builds (#1489)
* SMAGENT-1777: Fix code coverage builds

In a recent change we introduced the idea of build variants.  There's a
build variant to handle code coverage, but that build didn't include the
necessary options to generate code coverage info.

This change also uses the fault injection framework to test a previously
untested piece of code.  (I wanted a sample to show during a future
demo.)