SMAGENT-1789: Move dragent_logger out of dragent (#1499)
* SMAGENT-1789: Move dragent_logger out of dragent

We currently have a logger, `dragent_logger`, that is used by dragent.
We have a couple of other components (e.g., userspace-shared,
libsanalyzer) that ought not depend on the dragent component that might
like to use the logger.  This change pulls the logger out of `dragent`
and into `userspace-shared`.

We also have a logger in sysdig (`sinsp_logger`).  Both `dragent_logger`
and `sinsp_logger` were defined in files name `logger.h`.  Since we have
a flat include space, when you `#include "logger.h"`, who knows which
one you'll actually pick up?  Since this change is moving
`dragent_logger` out of `dragent`, I renamed it to `common_logger` and
put it in `common_logger.{h,cpp}`.  This resolves the ambiguity around
`logger.h`.