SMAGENT-1765: statsd_forwarder drops metrics (#1510)
* SMAGENT-1765: statsd_forwarder drops metrics

SMAGENT-1765 documents an issue observed in our production environment
where the agent drops some statsd messages when the `statsd:use_forward`
option is enabled.  SMAGENT-1797 documents an issue observed in our
staging2 environment where some statsd metrics from the backend get
corrupted somewhere between being generated and being displayed on the
Monitor UI.  After investigation, I observed that if the size of the
message that includes statsd data exceeded a certain length, we started
dropping metrics.  That prompted me to examine the code, and I found
that the `statsd_server` implementation was using a fixed-size 2048-byte
buffer to read data from the UDP socket.  If a client sent a UDP
datagram with more than 2048 bytes, this would:

* Cause us to drop any metrics that came after the 2048-byte mark
* Cause us to send a single corrupted statsd metric if the last metric
  was written across the 2048-byte mark.

This change updates `statsd_server` to dynamically resize the buffer
that it uses to hold the message to accommodate the entire message.  We
start off with a 1024-byte buffer, and resize it as needed.  With the
assumption that the rate at which metrics are generated is fairly
regular, this change does not try to scale-down the buffer size.