SMAGENT-1816: Don't assume std:: is available (#1519)
Sysdig pollutes the global namespace with everything in
`namespace `std. A ton of the agent code depends on that when
it really ought not. I'd like to eventually fix Sysdig to not
include `using namespace std` in header files, but that'll be
a big change that affects sysdig, oss-falco, and the agent. In
the mean time, this change fixes the agent code. There is some
risk of this creeping back into the agent before I can fix
sysdig, but if it does it'll be much less than it is now.

The approach:
* Scope resolve everything in `namespace std` in header files
* Either scope resolve everything in `namespace std` in implementation
  file, or explicitly add `using namespace std` to the implementation
  files if there were too many to fix.