SMAGENT-1814: Add unit tests for statsd_server (#1515)
* SMAGENT-1814: Add unit tests for statsd_server

This change adds unit tests for class `statsd_server`.  Since
`statsd_server` deals with both IPv4 and IPv6 addresses, this test will
not fail in some cases where it is unable to create sockets.  This
allows some flexibility in the environment in which the test runs.  I've
run the test locally on my box, which has both IPv4 and IPv6 and all
tests pass.  I've run the test in the build container, which has only
IPv4, and all the tests pass.