SMAGENT-1836: AppCheck Buckets Parsing Error (#1535)
* SMAGENT-1836: AppCheck Buckets Parsing Error

When we process an app check message of type "buckets", we expect all
bucket values to be convertable to UInt64.  If the bucket value is
negative, then the parser will throw an exception.

This change updates the code to omit the bucket value if parsing fails.
We also generate an error log message in that case.  If this class of
failure happens often, the log could be too verbose, but I hope this
doens't happen often.

There's another discussion around "aren't negative values valid?"  This
change doesn't try to address that issue, as it may impact the entire
product (at the very least the agent/protobuf/backend).

This change also adds a new unit test that for the `app_metric` class,
and uses that to validate the change.