SMAGENT-1824: statsd security limit limit failure (#1531)
* SMAGENT-1824: statsd security limit limit failure

When security is enabled, the agent is supposed to increase the
configured statsd limit by 100 to allow for statsd metrics that are
associated with the security feature.  Currently we're reading the
configuration value before the YAML is parsed, so the logic that
determines whether that limit increase should occur always says "no".

To fix this, we need to be able to read the _actual_ config value.
Rather than pass around the legacy `m_configuration` object, I decided
to move the security config to the new configuration framework.  This
change introduces namespace `security_config` that (1) stores all the
security-related config values, and (2) provides an API that enables
clients to read those values.  Here, by "security-related", I mean
"configuration values under the 'security' heading".  This change does
not, for example, move any Falco-related config options (although that
might be a nice follow-up change).

Most of this change is updating clients of the existing config to use
the new namespace.  The major changes are:

* The introduction of `security_config.{h,cpp,.ut.cpp}`
* The `statsd_emitter` family of classes no longer need to be passed a
  boolean indicating whether or not security is enabled; they now read
  the value via `security_config`.