Prefer thread with vtid=1 for cgroup selection. SMAGENT-1850
Our unstated assumption is that all processes in a container live
in the same cgroups so we can pick any one for reading stats
from cgroups.

It doesn't have to always be true (as shown by this issue, where
we introduce multiple cgroups for the agent itself), so if possible,
pick the thread with vtid=1 (container init), which (and again,
this is a heuristic) has a decent chance to belong to the parent
group of all the container cgroups.