[SSPROD-1835] Parametrized baselining config values
There's few configuration parameters for the baseliner, and they are
statically defined. We are currently not sure wether the default
values are good for our customer install base, therefore I'm proposing
to make it configurable through the usual agent config, in order to
gain flexibility. In a subsequent change, once we clearly identify
them, we should introduce hard limits that can never be surpassed even
through agent config.

The new agent config parameters are:
- falcobaseline.report_interval (nanoseconds)
- falcobaseline.autodisable_interval (nanosecond)
- falcobaseline.max_drops_full_buffer

The `report_interval' refers to the falco baseline emit cycle,
defaulted to 15 minutes.

The `autodisable_interval' refers to the time falco stays disabled,
upon it got auto-disabled because of perfomance issue on the host
(either sr being greater then 1, or full buffer being full).

The `max_drops_full_buffer' refers to the maxim number of event being
dropped in the falco baseliner full buffer.