SMAGENT-1886: Agent drops statsd messages (#1584)
* SMAGENT-1886: Agent drops statsd messages

In staging2 we observed that the agent was periodically dropping statsd
metrics from the collector.  After some investigation, I discovered that
when the agent was determining the set of containers to send to
`statsite_forwarder`, it would pick the first that it found in the
thread table as a reference thread.  If that thread was less than 10
seconds old, the analyzer would omit it; the rationale being that the
agent wanted to give the container time to start a statsd server if it
wanted.  In the presence of a process that creates new threads, this
behavior can cause the agent to omit containers if the newly-create
thread is selected as the representative thread.

This change updates the above-described behavior.  Now the agent informs
`statsite_forwarder` of all containers (not including the agent
container).  `statsite_forwarder` keeps track of the time at which it
first learned about each container.  Only after a configured amount of
time (default 0s) does `statsite_forwarder` create the `statsd_server`
for the container.