Make all `sinsp_container_info`s in the manager map const. SMAGENT-1892
This means that custom containers have to replace the whole container object
every time new metadata is found. Note that currently we have no logic
to determine if the metadata actually changed, so we replace the whole
object every time. Before doing something smarter, we'd have to check if
the added overhead to detect actual updates is smaller than the alloc/free
cost.