Ssprod 1874 fix container only policy scopes (#1595)
* Handle container-only scopes outside infra state

For any host-only activity, the container id for the syscall event is an
empty string, but the infrastructure state code does not know of an
entity with an empty container id.

There needs to be a shortcut like we already have for host-only activity
that examines the scope predicate and circumvents the infrastructure
state code for a container-only scope.

This fixes SSPROD-2164.

* New tests for {host,container}-only scopes

New automated tests that check for host-only and container-only scopes.

Each test has a policy looking for writes of /tmp/{second,third} and the
test does writes to that file both on a host and in a container. The
test then verifies that the only policy events were from the container
or host, respectively.

This required some extension to expected_policy_event to add a flag that
limits the events to containers, hosts, or either. The default is either.