[SMAGENT-1806] handle NaN values for raw prometheus metrics (#1607)
Before this change, if the prometheus scraping code saw a NaN value from
a scraped metric, the generated json response to app_check_proxy would
cause the jsoncpp library to error on the entire response.  (See the
'allow_nan' parameter to json.dumps that defaults to true.)

If a prometheus metric reports a NaN value, as can happen with
summaries:
https://github.com/prometheus/prometheus/issues/860
then we do need to send it to the backend. Pragmatically, these are a
pain to work with, but they are allowed values, so we need to make them
observable via our prometheus API.