Fix broken build caused by go 1.13 (#1661)
The go modules changes were tested in go 1.12, but by the time they were
merged, go 1.13 had been released. Modules changes in 1.13 broke the build
in a couple ways.

https://golang.org/doc/go1.13#version-validation
The old docker version we're using is specifically called out in the release
notes because it doesn't support modules correctly.

Also, some imports in kubecollect test stopped working with the local import
path and had to be updated to the path used in the replace directive.