SMAGENT-2028: Make type_config::get less ambiguous (#1667)
Currently, `type_config` has a `get()` method to return the
config value. That made sense when we didn't have the
`type_config_builder` that returns a smart pointer. With the
addition of the builder, you can have

```c++
c_config.get()
```
Which will either:
* Get you the value of the config object or
* Get you the pointer the smart pointer is holding

All depending on whether you used the builder or not. This makes it easy
to introduce hard-to-find bugs (and we have).

This change renames `type_config`'s `get()` API to `get_value()`.
This change also renames `type_config_builder`'s `get()` API to
`build()` (and similarly, `get_mutable()` to `build_mutable()`).

Unfortunately, we've been making config objects part of components'
public interface, so this change is more massive than I had hoped it
would be.