Ssprod 1590 use statsite config compliance metrics (#1675)
* Honor statsite cfg (port) when running comp tasks

When starting compliance tasks, pass along the configured statsd port
and send metrics to that port.

This allows configuring the agent with a statsd on an alternate port.

This fixes SSPROD-1590.

* Test for compliance + alternate statsd port

Add an automated test that verifies compliance metrics are sent to an
alternate statsd port if configured.

A new test class compliance_test_alternate_statsd_port derives from
compliance_test but uses port 8188 instead of 8125.

In the base class, use the configuration_manager to set the statsd port
as a part of setup. The original configuration is replaced when the test
is torn down.