[SSPROD-1879] Fixing baseliner reporting inconsistent container image name
Currently, the baseliner is reporting the m_image field from the
container_info. This field is not consistently reporting
human-readable image names, but instead the literal name upon the
execution of `docker run' (for instance). When a container
orchestrator like Kubernetes is involved, we don't get the actual
desired image name. Moving to use repo/tag/digest fields, to always
get the correct human-readable image name.

Signed-off-by: Lorenzo David <lorenzo.david@sysdig.com>