[SSPROD-2458] Provide a higher drop_upper_threshold when baselining
In its current implementation, the baseliner is auto-disabled once the
agent start to drop events. If the baseliner is enabled, we require a
higher drop_upper_threshold in order to properly collect fingerprints.

This change introduce a new constant for handling the default value
when baseliner is enabled (set to 30).

In case of a user provided drop_upper_threshold, we do not override,
but simply warn the user about the recommended value.

In this change, I have also corrected two misspelled occurrences of `Number'.

Signed-off-by: Lorenzo David <lorenzo.david@sysdig.com>