SMPROD-2494: Ignore legacy BASELINES message (#1697)
* SMPROD-2494: Ignore legacy BASELINES message

The `BASELINES` message is legacy; the agent can ignore the message.
This change introduces a `null_message_handler` that will drop any
messages that it receives.  It updates the initialization of the
connection manager to register a `null_message_handler` for the
`BASELINES` message.  This change deletes the old
`security_baselines_message_handler` message handler, and the
`security_baselines_loader` interface that the baselines message handler
used.