SSPROD-2402 - Agent hang after config update (#1696)
* SSPROD-2402 - Agent hang after config update

When the agent starts for the first time, it usually goes through this
sequence of events:

* Starts
* Receives a `CONFIG_DATA` message from the backend
* Updates config file based on message
* If the file changed (which it does the first time), then the agent
  restarts
* Starts again
* Receives a `CONFIG_DATA` message from the backend
* Updates config file based on message
* This time the file didn't change, so the agent continues execution

During startup, the security-related code does some of its on
initialization.  It is possible that that initialization is still in
progress when the agent receivies the `CONFIG_DATA` message that trigger
the reboot.  When that happens, the agent can deadlock between:

* One thread performing initialization
* Another thread blocks waiting for teardown to complete

To address this potential deadlock, this change postpones the
initialization of the `security_mgr` and `compliance_msg` components
until *after* the agent receives the second `CONFIG_DATA` (or until a
30s timeout expires).  The agent saves any `POLICIES`, `POLICIES_V2`, or
`COMP_CALENDAR` protobufs that it receives while waiting for the
timeout, and applies those changes after the security components are
initialized.