[SSPROD-2533] Auto-disable baseliner on high drop rate
The baseliner emit periodically (by default every 15 minutes)
fingerprints (an aggregate information regarding host and containers)
to the collector. In order to keep the quality of the collected data
high, the baseliner has two auto-disable mechanism. The rationals
behind these two mechanism is to ensure that (a) the baseliner doesn't
take up too much resources when the agent is overcharged and (b) in
order to keep the quality of the baselines high.

Currently, the baseliner can be turned off either because of sampling
ratio being higher then 1, or because the analyzer accounts for too
much dropped events.

This change improve the latter auto-disable mechanism, and introduce
the concept of drop rate.  The baseliner will keep internal statistics
regarding the number of total events and number of dropped events (due
to the buffer being full) across two emission cycle (the previously
mentioned 15 minutes by default). The user will be able to configure
using the `falcobaseline.max_drops_buffer_rate_percentage'
configuration variable the percentage upper threshold upon which the
baseliner will be disabled.

As usual, the baseliner will be auto-enabled (if all auto-disable
conditions are met) after `falcobaseline.autodisable_interval_ns' (by
default 30 minutes).

Signed-off-by: Lorenzo David <lorenzo.david@sysdig.com>