[SSPROD-2455] Adding support for profiling v1.2
This change is part of the profiling v1.2 milestone. The Secure
Profiling Worker will be the only client/consumer of this NATS stream.

The data flow follows closely what has been done for the Secure Audit
Stream data pipeline: a new type `SECURE_PROFILING_FINGERPRINT` has
been added in the protobuf. The agent forwards profiling fingerprint
messages to the collector on a separate protobuf message.

The payload containing the fingerprint is compressed to reduce the
network bandwith overhead.

This change outdates the previous implementation of the data emission
of `falco_baselines`, which was embedded as an optional field within
the `metrics` protobuf.

We defer to a new change the cleanup of the whole legacy baseline
implementation, which pre-dates the Secure Profiling project.

In conclusion, the name baselining and profiling has been used
interchangebly in a very confusing way. We also defer to a change a
more in-depth naming consolidation.

Signed-off-by: Lorenzo David <lorenzo.david@sysdig.com>