Add plumbing for compression method (#1805)
* Add plumbing for compression method

The 10s flush project will need the ability to set the compression
method on serialization. This work is in preparation for those changes,
and is intended to be forward-looking for when the backend rolls out
support for alternate compression methods (not just true or false).

The goal currently is no noticeable agent change.