Add addl info healthcheck tests (#1829)
* Run failed tests a second time with -v

In cases where some dragent tests fail and get run a second time, run
thoses tests again with "tests -v", which enables debug logs. This could
help reveal the source of the failed tests.

* Show event counts for hcheck/liveness test fails

In the healthcheck/liveness/readiness tests, we have a capture_stats
function that displays info on the number of events processed + dropped
that can be printed when an ASSERT_TRUE/etc fails.

Also use that function for the ASSERTS that verify that the healthcheck
command was seen.

* Set sinsp log cb for trace tests

The tracefile tests create a separate inspector so it can dump the
events to a capture file that is re-read. To log everything with -v,
set the log callback for that inspector to the common logger.