SMAGENT-1967: Experian sinsp_workers stall (v2) (#1853)
* SMAGENT-1967: Experian sinsp_workers stall (v2)

Experian has a problem where the agent starts, it gets behind in
processing system call events early, and restarts itself in a loop.
Our best understanding at this time is that after startup the agent
is busy with initialization tasks, and that that can contribute to
it getting behind in processing system call events.

This change introduces a new configuration option,
`inspector_start_delay_s` that enables customers to delay starting the
inspector for a given number of seconds.  With this, customers can
configure the agent in such a way that it starts, delays inspector
startup until it completes its initialization tasks, then starts system
call processing.  The option defaults to 1, and 1 is the minimum value.

This change extends the first approach by running a proc scan
immediately after restarting the capture.  This should enable the agent
to fill in gaps for any processes that were started between the time
that the capture was stopped and restarted.

This change also introduces a new utility component, `run_once_after`.
Clients periodically call `run()` on instances of this class, and the instance
will invoke the given function once after a pre-defined period of time
passes.