Make healthcheck tests tolerant of lookup delays (#1857)
In jenkins, we noticed that it some cases it can can take a really long
time (as much as 1.2 seconds) for the async lookup of container
information to complete. This is a problem for the healthcheck tests as
the container itself only runs for 1 second.

Make the tests more tolerant by running the top level container for 10
seconds instead of 1 second. The tests already short-cut and kill the
container once the conditions for the test are met (observing both the
initial process and/or the specific healthcheck/liveness/readiness
processes), so the common case should still be fast.

One additional change is that we don't require that when seeing a
healthcheck process that it always has a healthcheck category. This
could happen if the first healthcheck process is observed while
the (long) async lookup is still in progress. We pass the test if we
eventually see the healthcheck process with the appropriate categrory
for the test.