SMAGENT-2249: Kafka Tap: Aggr network audit data (#1884)
* SMAGENT-2249: Kafka Tap: Aggr network audit data

As part of SMAGENT-1981 we delivered to Goldman a feature that included
network audit data once per minute.  At the time the requirement was
that the data must include only the last 1 second's information (i.e.,
once per minute we send 1 second of network audit data).  When that
feature was delivered, they asked for an extention of that work to send
aggregated information once per minute.  This change implements that
extension.

This change introduces:
* `audit_tap_network_connection` -- a class that models an audit tap
   network connection "key" (values that are used to uniquely identify
   a network connection).  Instances of this class are suitable as keys
   in a `std::unordered_map`.

* `audit_tap_connection_info` -- a class that models that data we
  aggregate for each network connection.

* `audit_tap_connection_aggregator` -- a class that maintains the
  aggregation over the aggregation period (60s by default).

* Unit tests for the above three components

This change updates the existing `audit_tap` component to use the
newly-introduced `audit_tap_connection_aggregator` to save and
eventually populate the aggregated data.