Don't symlink gcc to gcc-4.8/4.9 [SMAGENT-2312] (#1940)
A recent change
https://github.com/draios/agent/commit/ddaaceab15d0c5f9471595c94064c89d5204cfd9
improved the docker entrypoint to check for the specific gcc version
used by the kernel and symlink that version to /usr/bin/gcc. However, if
the gcc version happened to be gcc-4.8 or gcc-4.9, that would conflict
with an older change in the agent's Dockerfile that symlinks
/usr/bin/gcc to gcc-4.8/4.9.

The obvious fix seems to be to set up the symlinks only in
docker-entrypoint.sh and not in the Dockerfile, however legacy Debian kernels
have `gcc-4.8` and `gcc-4.9` hardcoded in their Makefiles. This means
that we need to have a compiler under these names and we can't point
the `gcc` symlink at them. Instead, only consider gcc versions 5-9
as candidates for symlinking /usr/bin/gcc to.