Separate debug/warning level messages from config file
We effectively never log warnings coming from the config parser
because they use the debug level.

LOG_DEBUG is probably appropriate for missing config files
(we have a lot of them and at least some won't exist) but not really
for any other messages that are:
* caused by user configuration
* not critical but still important

Introduce a separate list of debug-level messages, move
the "missing config file" there and log debugs/warnings/errors
with the appropriate levels.