Ssprod 2830 update kube bench (#1939)
* Pass comp task param "benchmark" as kube-bench arg

If a compliance task has a parameter "benchmark", add the --benchmark
argument to the kube-bench command line, passing the value as the
arg. For example, if the param "benchmark" -> "cis-1.4" is provided, run
kube-bench with "--benchmark cis-1.4".

* Update kube-bench to 0.2.3

This has additional fixes for redhat hardening guide for minishift.

Also use a more recent kubectl.

* Set result ResultSchema to benchmark when present

If the task specifies a benchmark version like rh-0.7 for openshift
hardening guide, set the SchemaVersion to that benchmark version. This
allows the backend to properly apply the right labels to the result
sections + tests.

* Consider INFO kube-bench tests as PASS

When running kube-bench with the openshift hardening guide, some tests
have have an INFO status. This wasn't being handled properly by the
scraper, which previously only knew about PASS/FAIL/WARN. The
consequence was incorrect overall counts and the INFO individual tests
being characterized as FAILs.

More properly handle them as PASS.

This fixes SSPROD-3314.