Ssprod 3508 add container kill action (#1989)
* Agent changes to support container kill actions

Following the protobuf changes in
https://github.com/draios/protorepo/pull/46, support a separate list of
v2actions that allows for more flexible extensions for adding future
actions.

This agent supports STOP/PAUSE/CAPTURE in either of the actions or
v2actions arrays, and additionally KILL in the v2actions array. The
action result goes in the corresponding action_results/v2action_results
part of the policy event. For both arrays, if the action is unknown set
the result to complete with successful=false and an error message. The
policy event will be sent when every action and v2 action has a
result/v2result.

Perform_container_action() needs to be a bit more generic to pass
around either a action_result or v2action_result. Add a wrapper
class that handles setting successful/errmsg/token and pass that to
perform_container_action().

Move the details of requesting a capture to a standalone method
perform_capture_action() like perform_container_action(), which can work
with either v2actions or actions.

Define a KILL command between the agent and cointerface, corresponding
to ACTION_KILL.

For docker, use cli.ContainerKill to kill the container. For cri, still
use StopContainer but set the timeout to 0 to immediately stop the
container.

* Break out capture job enqueues to standalone class

Break out the part of the capture job handler that accepts requests into
its own virtual void class. New class capture_job_queue_handler has the
types for job requests and the single method queue_job_request (virtual
void).

This way we can stub out the capture job requests in unit tests to its
own class and add tests for capture actions.

* Tests for cont. kill, capture policy actions

Add tests for container kill policy actions as well as capture actions.

Each of stop/capture needs test variants that can work with either
actions or v2actions. They all work on executed processes rules as
follows:

 - actions stop: gzip
 - v2actions stop: bzip2
 - actions capture: bunzip2
 - v2actions capture: bzcat
 - v2actions kill: lzcat

The stop tests also have variants that work with fake cri containers
launched by test_helper.

There are shared functions that work for either actions or v2actions,
but otherwise share the same test startup/teardown.

Move validation of actions into expected_policy_result and use that to
validate actions instead of the old stop validation that iterated over
the policy events protobuf directly.

Also, change the docker commands to ensure that the docker run command
has a non-zero result. We also change this for stop--as it was, the
docker container was actually exiting with 0 not because it was killed,
but because the container finished its (short) sleep before the up-to-30
second delay for the stop completed.

Utilize the newly separated capture_job_queue_handler interface to
create a fake capture job handler that simply queues requests. Add tests
that have capture actions, and expect to find that the test capture job
queue handler has gotten a dump request from the security manager after
receiving the policy event.

Also add a test for unknown actions. The test ensures that the policy
event is still generated, but with an unsuccessful action and an error
message stating that the action is unknown. As the actions enum is
extended, the new values will get mapped to 0 so this test validates
that error path.