[SMPROD-3121] Change draios.proto handling to use configed go_package (#2051)
In the main PR for this change (draios/metadata-service#3) we’re creating a new KubernetesStateRequest message which wraps the existing draiosproto.k8s_state. In practice, this means that we want to autogenerate golang grpc code for a new proto which references an existing proto in a separate directory. This was very flaky and small changes would lead to weird errors.

The internet talked about this problem and the conclusion is that protogen can’t reliably detect the correct package unless we use the go_package option in our protobuf (draios/protorepo#54). Future versions of protogen will require that this option is set. Adding that option broke the agent code which led a PR to fix the agent (#2051).

Another positive of the go_package option is that we can just generate the code to a folder then change our GOPATH and we're confident about what we need to import.