Ssprod 3786 falco exact rule names only (#2058)
* Only match exact rule names for falco

When adding falco rules to a ruleset (which are 1-1 with policies), only
match exact rule names. This prevents problems like SSPROD-3786, where a
policy with rule "...Read sensitive file from Web Server" was also
matching a rule "...[GS] Read sensitive file from Web Server - v3", as
the first rule was a substring of the second.

* Add tests for falco exact rule name matches

All that's required is to add a policy with rule_names that is a
substring of an existing rule name. It should *not* change the number of
matching events for existing tests.

* Use the master branch of oss-falco

Upstream falco switched to using master for day-to-day updates, and that
branch is being synced to oss-falco, so we should use that branch for
agent builds.