Allow master to be the target of a sync (#2082)
Evidently, when GitSCM checks out the master branch, it does so to a
detached HEAD and not a real branch. In turn, that prevents checking out
the master branch later.

To fix this, add the LocalBranch extension and additionally check out
the master branch before adding the upstream remote.

Also clean up the workspace so each time the job starts from scratch.

Finally, there's no need to check out the agent directory. The
jenkinsfile is checked out separately, and the downstream jobs do their
own checkouts.