[SMAGENT-2398] Force ack of the requested sampling ratio, not acked
When we switch from sr=2 to sr=1 in the non-dropping period (first half
of every second), the driver does not emit a DROP event confirming
the sampling ratio change. We work around this by assuming the switch
happened but we set the sampling ratio to the already acked value,
not the requested one.