Ssprod 4514 syscall matchitems false (#2193)
* Only consider enter events in syscall rules

When loading a rule and building the lists of event type to match info
struct, only the enter events are used.

For matchItems=true lists, this is okay, as the exit events just end up
getting skipped and the enter events get compared in the lists.

For matchItems=false lists, though, this results in false positives, as
the exit event ends up not matching the items in the list, simply
because the list didn't have any exit events.

This fixes SSPROD-4514.

* Add additional logging for syscall rule loads

Note when the object is resent and each event name/number and whether it
matches or not.

* Unit test for matchItems=false + syscalls

Unit test that verifies the fix for SSPROD-4514. The test has a rule +
policy scoped to a container that allows all syscalls you would expect
to see in a busybox container that reads files, sleeps, etc. The only
allowed syscall/event type is procexit.

The test creates a container, and verifies that the only policy events
are for the whitelist policy for procexit events.