Ssprod 3873 policy scope cluster name (#2244)
* Add kubernetes.cluster.name as a host tag if set

This should allow for scope matching against the cluster name. Hosts
will have the tag, and I believe containers use the host tags as well.

* Tests for scope matching against cluster name

A new test class "security_policies_v2_test_cluster_name" optionally
sets the cluster name to my-cluster.

A new policy is scoped to kubernetes.cluster.name=my-cluster. 4 tests
verify that for host & container activity, that an agent configured with
k8s_cluster_name=my-cluster detect activity for the policy and an agent
configured with the default k8s cluster name do not.

Also change how memdump/secure are enabled so the config makes it all
the way to the analyzer on_capture_start(). This makes sure that infra
state has the right configured cluster name.

Also use the file-based approach to set the cri socket, which is more
resilient than scoped_configuration.

The test does have to wait for the socket to exist, though--otherwise
the config module will clear the value.

* Add addl docker/cri command logging.