[SMAGENT-2772] Ensure pod_restart_rate is always increasing
We had customers experiencing spikes in the pod_restart_rate metric.
This happens if pod_restart_count is reset to 0. Passing from the
actual value to 0 gives a negative rate that casted to an unsigned int
returns a  ~2^32 value. Considering that when a reset is done no
restart happens, we handle this corner case passing to the BE the max
between the calculated rate and 0