[SSPROD-5535] Fixing netsec issues when audit stream is disabled (#2296)
This change allows netsec to work in the context of audit stream and
audit tap to be disabled. With this fix, the connection tinfo are
properly stored when netsec is enabled, and therefore allows netsec to
properly lookup container metadata.

Note that connections that are not related to a container are
discarded by netsec. This issue was causing netsec not to track and
report any connection, as no connections were actually being marked as
belonging to a container.

Signed-off-by: Lorenzo David <lorenzo.david@sysdig.com>