[SMAGENT-2051] 10.8.0 jenkins update (#2315)
Several changes to our release jenkins jobs to deal with new (stuff) that needs to happen as well as address some issues from previous releases

we now have a mirror of our released containers in quay. The promote job is now required to take those dependencies and mirror them to quay. This code should work and has been "test"ed
due to the increasing complexity of the promote job, there is an increasing chance of intermittent failures. This occurred with 10.7, and it was a pain to effectively finish the promote manually. We now use the jenkins retry mechanism to allow these stages to be retried to deal with intermittent failures. We should probably add some sort of delay...but baby steps. In theory this has also been tested
I finally figured out how to dynamically send messages to the release channel, so updated some notifications. Namely moving some from release->release channel, as well as removing the docs channel, since the relevant docs individual should be in the release channel anyway. So it just reduces some spam for them. The main idea has been tested, but obviously each message cannot be
we produce a build for GS that we sometimes give them for beta testing before a release is out. This is now built automatically after the RC is built. The job has been tested, but the invocation has not.